refactor(api): updated to calculate summary stats on the server side, still not included in response
the inclusion of the summary stats in the api resp will require some changes in the react app so I
need to perform those changes at the same time. Also need to add typing to the python code so it's
more readable and understandable, esp w.r.t. lsp working on it in editor